requirements.txt
streamlit
praw
spacy
textblob
matplotlib
wordcloud
python-dotenv
